refactor: Image 관련 API 정리
