+++ src/api/image/type.ts
@@ -1,0 +1,6 @@
+import { paths } from '@/__generated__/schema2';
+
+export type GetPresignedUrl = {
+  request: paths['/meeting/v2/presigned-url']['get']['parameters']['query'];
+  response: paths['/meeting/v2/presigned-url']['get']['responses']['200']['content']['application/json;charset=UTF-8'];
+};
